+++ src/converse-minimize.js
@@ -505,7 +505,7 @@
                 }
             });
 
-            var renderMinimizeButton = function (evt, view) {
+            var renderMinimizeButton = function (view) {
                 // Inserts a "minimize" button in the chatview's header
                 var $el = view.$el.find('.toggle-chatbox-button');
                 var $new_el = converse.templates.chatbox_minimize(
@@ -519,7 +519,7 @@
             };
             converse.on('chatBoxOpened', renderMinimizeButton);
 
-            converse.on('controlBoxOpened', function (evt, chatbox) {
+            converse.on('controlBoxOpened', function (chatbox) {
                 // Wrapped in anon method because at scan time, chatboxviews
                 // attr not set yet.
                 if (converse.connection.connected) {