+++ docs/CHANGES.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 2.0.5
+## 3.0.0 (Unreleased)
+- Breaking change: Callbacks for `converse.on` now no longer receive an event
+  object as first parameter. [jcbrand]
+
+## 2.0.5 (Unreleased)
 - #743, #751, #753 Update to Strophe 1.2.12. SASL-EXTERNAL now has reduced priority, so it won't
   be prioritized above other auth mechanisms. [jcbrand]
 - #755: create composer.json to add this project in packagist.org [fabiomontefuscolo]